Refactored functionality into their own modules
