+++ app/skywrapper/enums.py
@@ -1,0 +1,8 @@
+from __future__ import unicode_literals, print_function, division, absolute_import
+
+from enum import Enum
+
+
+class Direction(Enum):
+    inbound = 'Inbound'
+    outbound = 'Outbound'
